% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{name_cleaner}
\alias{name_cleaner}
\title{Column renamer}
\usage{
name_cleaner(x)
}
\arguments{
\item{x}{Character vector of column names or data.frame/data.table object}
}
\value{
The same names as input, but lowercase with underscores instead of
  spaces (multiple spaces are reduced to single)
}
\description{
Strip out the punctuation and white space from imported colnames so they can be
  tab autocompleted more easily
}
\examples{
raw_names <- c("  The first column  ", "_second col _ ")
name_cleaner(raw_names)
# [1] "the_first_column" "second_col"

raw_names <- c("A_", "_a", "  A _")
name_cleaner(raw_names)
# [1] "a_1" "a_2" "a_3"

}
